refactor: 컨트롤러 의도 명확히 분리
