+++ src/main/java/roomescape/controller/ReservationQueryController.java
@@ -1,10 +1,24 @@
 package roomescape.controller;
 
+import java.util.List;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.ResponseBody;
+import roomescape.domain.Reservation;
+import roomescape.service.ReservationService;
 
 @Controller
 public class ReservationQueryController {
+
+    private final ReservationService reservationService;
+
+    @Autowired
+    public ReservationQueryController(ReservationService reservationService) {
+        this.reservationService = reservationService;
+    }
 
     @GetMapping("")
     public String home() {
@@ -19,5 +33,24 @@ public class ReservationQueryController {
     @GetMapping("admin/reservation")
     public String reservation() {
         return "admin/reservation-legacy";
+    }
+
+    @GetMapping("admin/time")
+    public String time() {
+        return "admin/time";
+    }
+
+    @GetMapping("reservations")
+    @ResponseBody
+    public ResponseEntity<List<Reservation>> readReservations() {
+        List<Reservation> reservations = reservationService.readAll();
+        return ResponseEntity.ok(reservations);
+    }
+
+    @GetMapping("reservations/{reservationId}")
+    @ResponseBody
+    public ResponseEntity<Reservation> readReservation(@PathVariable("reservationId") Long id) {
+        Reservation reservation = reservationService.readOne(id);
+        return ResponseEntity.ok(reservation);
     }
 }
